pyproject.toml
[project]
name = "olimpiadas-prog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-anymail>=13.0",
    "django-extensions>=4.1",
    "django-ninja>=1.4.3",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "uvicorn>=0.34.3",
    "whitenoise>=6.9.0",
]
